% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_form.R
\name{app_form}
\alias{app_form}
\title{App Form}
\usage{
app_form(id, class, button_id, ...)
}
\arguments{
\item{id}{The html-id for the form.}

\item{class}{The css classes to be applied to the form tag}

\item{button_id}{The html-id for the button}

\item{...}{Any  number of html tags to be inserted into the form body.}
}
\value{
html form tag
}
\description{
This function creates the html from container.
}
