% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appifyr.R
\docType{package}
\name{appifyr-package}
\alias{appifyr}
\alias{appifyr-package}
\title{appifyr: Create Web Applications from R Functions}
\description{
The appifyr package enables R users to create interactive web applications
using OpenCPU as a backend. It provides a bridge between R's analytical
capabilities and web interfaces by automatically generating HTML forms
for function parameters, connecting these inputs to OpenCPU API calls,
and displaying the function outputs.

To learn more about appifyr, start with the vignettes:
`browseVignettes(package = "appifyr")`
}
\section{Key Functions}{

\itemize{
  \item \code{\link{appify}}: Create a web app from an R function
  \item \code{\link{inp_text}}: Create a text input element
  \item \code{\link{inp_number}}: Create a numeric input element
  \item \code{\link{inp_dropdown}}: Create a dropdown/select input element
  \item \code{\link{build_app}}: Build and install the application
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/retowyss/appifyr}
  \item Report bugs at \url{https://github.com/retowyss/appifyr/issues}
}

}
\author{
\strong{Maintainer}: Reto Wyss \email{reto.wyss.ch@gmail.com} [copyright holder]

Other contributors:
\itemize{
  \item Michael Rustler \email{michael.rustler@kompetenz-wasser.de} (\href{https://orcid.org/0000-0003-0647-7726}{ORCID}) [contributor]
}

}
\keyword{internal}
