% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args_to_json.R
\name{arg_to_json}
\alias{arg_to_json}
\title{Argument To JSON}
\usage{
arg_to_json(an, id)
}
\arguments{
\item{an}{argument name, e.g. "n"}

\item{id}{html id}
}
\value{
json body code,
}
\description{
Argument To JSON
}
\examples{

arg_to_json("n", "xGkrZCuuI")

}
