% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_elements.R
\name{atomic_to_df}
\alias{atomic_to_df}
\title{Convert List Of Values To Dataframe}
\usage{
atomic_to_df(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a dataframe
}
\description{
Convert List Of Values To Dataframe
}
