% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_form.R
\name{canvas}
\alias{canvas}
\title{Canvas}
\usage{
canvas(id, height)
}
\arguments{
\item{id}{The html-id for the canvas}

\item{height}{The height for the plot}
}
\value{
target div for the plot
}
\description{
The canvas is the container for the plot your function returns.
}
