% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_inputs}
\alias{check_inputs}
\title{Check Arguments}
\usage{
check_inputs(f, input_names)
}
\arguments{
\item{f}{A function}

\item{input_names}{Vector of input names}
}
\description{
Check the actual inputs of a function against the provided
    inputs for the form inputs. If inputs do not match then a
    helpful message will be displayed. Returns invisible otherwise.
}
\examples{

# This passes
check_inputs(rnorm, c("n", "mean", "sd"))
check_inputs(rnorm, c("n"))

# This fails
check_inputs(rnorm, c("n", "mean", "q"))

}
