% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_id}
\alias{create_id}
\title{Create a Random Id}
\usage{
create_id(n = 10, prefix = "")
}
\arguments{
\item{n}{Length of the id (default: 10)}

\item{prefix}{Optional prefix to add to the random ID}
}
\value{
A random id string
}
\description{
Creates a random identifier from a-zA-Z. This function
  is used internally to generate unique identifiers for HTML elements.
}
\examples{
\dontrun{
create_id()
create_id(15)
create_id(8, prefix = "app_")
}

}
