% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_inputs.R
\name{dropdown}
\alias{dropdown}
\title{Dropdown}
\usage{
dropdown(label = "", width = 4, opts)
}
\arguments{
\item{label}{The label to display above the input field}

\item{width}{The width in bootstrap grid units (defaul 4)}

\item{opts}{Either a list of (key1 = value1, ..., keyn = valuen) or a two
column dateframe with column names "key" and "value"}
}
\value{
function(id) that will create the from dropdown with the specified id
}
\description{
Dropdown
}
