% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_inputs.R
\name{dropdown_color}
\alias{dropdown_color}
\title{Dropdown Colors}
\usage{
dropdown_color(label = "", width = 4)
}
\arguments{
\item{label}{The label to display above the input field}

\item{width}{The width in bootstrap grid units (defaul 4)}
}
\value{
function(id) that will create the from color-dropdown with the specified id
}
\description{
Dropdown Colors
}
