% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ensure_id_postfix}
\alias{ensure_id_postfix}
\title{Ensure Id Postfix}
\usage{
ensure_id_postfix(postfix)
}
\arguments{
\item{postfix}{Character string when set by user, else NULL}
}
\value{
The postfix or a random postfix
}
\description{
Ensures that an ID postfix is available, either using
  the provided value or generating a random one.
}
\examples{
\dontrun{
# Returns a random ID
ensure_id_postfix(NULL)

# Returns "myid"
ensure_id_postfix("myid")
}

}
