% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_app.R
\name{extract_pkg_names}
\alias{extract_pkg_names}
\title{Extract the package name}
\usage{
extract_pkg_names(pkg)
}
\arguments{
\item{pkg}{"library(pkg)" or "require(pkg)"}
}
\value{
a package name, e.g. dplyr
}
\description{
Extract the package name
}
