% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_app.R
\name{extract_r_functions}
\alias{extract_r_functions}
\title{Extract R functions from Rmd}
\usage{
extract_r_functions(rmd)
}
\arguments{
\item{rmd}{Rmd as character}
}
\value{
list of R functions
}
\description{
Extract R functions from Rmd
}
