% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_elements.R
\name{factor_to_chr}
\alias{factor_to_chr}
\title{Convert Factor To Character}
\usage{
factor_to_chr(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a vector
}
\description{
Convert Factor To Character
}
\examples{
appifyr:::factor_to_chr(factor(c("A", "B")))
appifyr:::factor_to_chr(factor(1:12))

}
