% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_elements.R
\name{fg_text}
\alias{fg_text}
\title{Translate Form Group List to HTML}
\usage{
fg_text(g, id)
}
\arguments{
\item{g}{a form group list}

\item{id}{an id}
}
\value{
html
}
\description{
Translate Form Group List to HTML
}
