% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_elements.R
\name{fg_to_html}
\alias{fg_to_html}
\title{Form Group To HTML}
\usage{
fg_to_html(g, id, theme = "bootstrap4")
}
\arguments{
\item{g}{a form group list}

\item{id}{an id}

\item{theme}{CSS framework theme to use}
}
\value{
html
}
\description{
Form Group To HTML
}
