% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_elements.R
\name{html_container_form}
\alias{html_container_form}
\title{App Form}
\usage{
html_container_form(id, class, button_id, ..., theme = "bootstrap4")
}
\arguments{
\item{id}{The html-id for the form.}

\item{class}{The css classes to be applied to the form tag}

\item{button_id}{The html-id for the button}

\item{...}{Any number of html tags to be inserted into the form body.}

\item{theme}{The CSS framework theme to use}
}
\value{
html form tag
}
\description{
This function creates the html form container.
}
