% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_elements.R
\name{inp_dropdown}
\alias{inp_dropdown}
\title{Dropdown Form Group}
\usage{
inp_dropdown(kv, label = NULL, width = 4, ...)
}
\arguments{
\item{kv}{a vector, a list or a dataframe}

\item{label}{label for the form input - this will default to display the
name of the variable}

\item{width}{the form-group's width in bootstrap grid units}

\item{...}{additional arguments passed to the form group parser}
}
\value{
a list
}
\description{
Dropdown Form Group
}
\examples{
inp_dropdown(c(1, 3, 7, 42))
inp_dropdown(names(iris), label = "Select Iris Variable")
kv <- data.frame(key = c("A", "B"), value = c(1, 2))
inp_dropdown(kv)
inp_dropdown(LETTERS, label = "Letters", width = 2)
inp_dropdown(list(n = 1, m = 2, k = 4))
}
