% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_elements.R
\name{inp_number}
\alias{inp_number}
\title{Number Input Form Group}
\usage{
inp_number(from = 0, to = 100, label = NULL, width = 4, ...)
}
\arguments{
\item{from}{a number}

\item{to}{a number}

\item{label}{label for the form input - this will default to display the
name of the variable}

\item{width}{the form-group's width in bootstrap grid units}

\item{...}{additional arguments passed to the form group parser}
}
\value{
a list
}
\description{
Number Input Form Group
}
\examples{
inp_number()
inp_number(from = 0, to = 1e9)
inp_number(from = 1, to = 7, label = "This Is A Survey")

}
