% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_elements.R
\name{inp_text}
\alias{inp_text}
\title{Text Form Group}
\usage{
inp_text(label = NULL, width = 4, ...)
}
\arguments{
\item{label}{label for the form input - this will default to display the
name of the variable}

\item{width}{the form-group's width in bootstrap grid units}

\item{...}{additional arguments passed to the form group parser}
}
\value{
a list
}
\description{
Text Form Group
}
\examples{
inp_text()
inp_text(label = "Plot Title")
inp_text(width = 6)
inp_text(label = "Description", width = 12)
}
