% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_.R
\name{js_rplot}
\alias{js_rplot}
\title{Make R-Plot JavaScript}
\usage{
js_rplot(canvas, button, r_function, args)
}
\arguments{
\item{canvas}{html-id of canvas}

\item{button}{html-id of submit button}

\item{r_function}{R function (as character)}

\item{args}{list of (argname = html-id)}
}
\value{
html script tag
}
\description{
Creates Javascript that handles the communication between
    OpenCPU and the website/webapp.
}
