% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_app.R
\name{new_app}
\alias{new_app}
\title{New App}
\usage{
new_app(template = "website")
}
\arguments{
\item{template}{template to build the app with (default: website)}
}
\value{
project directory is set up
}
\description{
This function should be called when starting a new project
    with appifyr. The default template is website and it's the only one
    currently supported. The templates files reside in the newly created
    app directory. They are standard Rmarkdown website components.
}
