% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_elements.R
\name{print.appifyr_app}
\alias{print.appifyr_app}
\title{Print method for appifyr_app objects}
\usage{
\method{print}{appifyr_app}(x, ...)
}
\arguments{
\item{x}{An appifyr_app object}

\item{...}{Additional arguments (not used)}
}
\value{
The app object invisibly
}
\description{
Print method for appifyr_app objects
}
