% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_elements.R
\name{set_label}
\alias{set_label}
\title{Set The Form Group Label}
\usage{
set_label(inps)
}
\arguments{
\item{inps}{a list of inputs}
}
\value{
form group list with labels
}
\description{
Set The Form Group Label
}
