% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_elements.R
\name{appify}
\alias{appify}
\title{Create App From Function and Inputs}
\usage{
appify(f, inps, out = "plot", id = NULL, theme = "bootstrap4")
}
\arguments{
\item{f}{a function as a string}

\item{inps}{a list of inputs, see \code{?inp_text}, \code{?inp_number}
and \code{?inp_dropdown}}

\item{out}{the output type (currently the only valid value is \code{"plot"},
with \code{"table"} planned for future releases)}

\item{id}{html identifier, when \code{NULL} the id will be generated
randomly. If you set the id manually, make sure it is unique.}

\item{theme}{CSS framework to use. Currently only "bootstrap4" is supported.}
}
\value{
An object of class "appifyr_app" containing the HTML components
  that can be included in R Markdown documents or Shiny applications.
}
\description{
Create App From Function and Inputs
}
\examples{
# Simple histogram function
# hist_fun <- function(n, bins) {
#   hist(rnorm(n), breaks = bins)
# }
# 
# Create app with numeric inputs
app <- appify(
  f = "hist_fun", 
  inps = list(
    n = inp_number(from = 10, to = 1000, label = "Sample Size"),
    bins = inp_number(from = 5, to = 30, label = "Number of Bins")
  )
)

}
