% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appify.R
\name{appify}
\alias{appify}
\title{Appify}
\usage{
appify(f, inputs = list(), id_postfix = NULL, input_check = TRUE,
  plot_height = 600, position_flip = FALSE)
}
\arguments{
\item{f}{A function from your package or written in your Rmd document.}

\item{inputs}{Inputs to your function for which a form will be created.}

\item{id_postfix}{Advanced users may wish to set the html-id postfix. The
default value is NULL. appify will then create a random postfix.}

\item{input_check}{Should inputs be checked against function inputs
(default = `TRUE`)?}

\item{plot_height}{Pixel value for the plot height. The default value is
600. So, the app including the form will comfortably fit on a
laptop screen.}

\item{position_flip}{By default (FALSE) the plot will be placed below the
input form.}
}
\value{
an app - including html and javascript
}
\description{
Appify
}
