% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_elements.R
\name{appify}
\alias{appify}
\title{Create App From Function and Inputs}
\usage{
appify(f, inps, id = NULL)
}
\arguments{
\item{f}{a function}

\item{inps}{a set of inputs}

\item{id}{html identifier}
}
\value{
html
}
\description{
Create App From Function and Inputs
}
