% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appifyr.R
\docType{package}
\name{appifyr-package}
\alias{appifyr}
\alias{appifyr-package}
\title{appifyr: appifyr: Write R Get Apps!}
\description{
To learn more about appifyr, start with the vignettes:
`browseVignettes(package = "appifyr")`
}
\author{
\strong{Maintainer}: Reto Wyss \email{reto.wyss.ch@gmail.com} [copyright holder]

}
\keyword{internal}
