% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_elements.R
\name{args_to_json}
\alias{args_to_json}
\title{Arguments To JSON}
\usage{
args_to_json(ans, id)
}
\arguments{
\item{ans}{argument names}

\item{id}{html identifier}
}
\value{
json body code
}
\description{
Arguments To JSON
}
