% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_app.R
\name{build_app}
\alias{build_app}
\title{Build App}
\usage{
build_app(app_dir = "app/website")
}
\arguments{
\item{app_dir}{path to the app (default: app/website)}
}
\value{
invisible
}
\description{
Build App
}
