% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_id}
\alias{create_id}
\title{Create a Random Id}
\usage{
create_id(n = 32)
}
\arguments{
\item{n}{Length of the id (default: 32)}
}
\value{
a random id
}
\description{
Creates a random identifier from a-zA-Z.
}
