% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ensure_id_postfix}
\alias{ensure_id_postfix}
\title{Ensure Id Postfix}
\usage{
ensure_id_postfix(postfix)
}
\arguments{
\item{postfix}{charchter when set by user else NULL}
}
\value{
the postfix or a random postfix
}
\description{
Ensure Id Postfix
}
