% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_elements.R
\name{html_container_output}
\alias{html_container_output}
\title{Container for Output}
\usage{
html_container_output(id, height)
}
\arguments{
\item{id}{The html-id for the ouput container}

\item{height}{The height for the plot}
}
\value{
target div for the plot
}
\description{
TThe ouput container is the target for the r function result.
}
