% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_io.R
\name{js_rplot}
\alias{js_rplot}
\title{Make R-Plot JavaScript}
\usage{
js_rplot(canvas, button, f_name, args)
}
\arguments{
\item{canvas}{html-id of canvas}

\item{button}{html-id of submit button}

\item{f_name}{R function (as character)}

\item{args}{list of (argname = html-id)}
}
\value{
html script tag
}
\description{
Creates Javascript that handles the communication between
    OpenCPU and the website/webapp.
}
\examples{

js_rplot(
 canvas = "canvas-id",
 button = "button-id",
 f_name = "function-name",
 args = list(A = "A-id", B = "B-id")
)

}
