% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kvs}
\alias{kvs}
\title{A list with key value pairs to be used with dropdowns.}
\format{A list with key value pairs to be used with dropdowns.}
\usage{
kvs
}
\description{
A list with key value pairs to be used with dropdowns.
}
\keyword{datasets}
