% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_elements.R
\name{list_to_df}
\alias{list_to_df}
\title{Convert Key Value List To Dataframe}
\usage{
list_to_df(x)
}
\arguments{
\item{x}{a list}
}
\value{
a dataframe
}
\description{
Convert Key Value List To Dataframe
}
