% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_inputs.R
\name{number}
\alias{number}
\title{Number Input}
\usage{
number(label = "", width = 4, from = 0, to = 100)
}
\arguments{
\item{label}{Input label}

\item{from}{The minimum value to accept as an input}

\item{to}{The maximum value to accept as an input}
}
\value{
html form input of type number
}
\description{
Number Input
}
