% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_elements.R
\name{rplot}
\alias{rplot}
\title{The R-Plot Widget}
\usage{
rplot(id, rf, json, err = "Error: ")
}
\arguments{
\item{id}{id of the corresponding html elements}

\item{rf}{name of the R function}

\item{json}{inner json of arguments to the R function}

\item{err}{error message for the javascript alert}
}
\value{
java script code
}
\description{
This function fills the rplot.js glue-partial. The rplot javascript
    function is part of the OpenCPU javascript library.
}
\examples{
appifyr:::rplot("xGkrZCuuI", "rnorm", "n : 5")
appifyr:::rplot("xGkrZCuuI", "rnorm", "n : 1000, mean : 100, sd : 15")

}
